requirements.txt
adaption
accelerate
datasets
evaluate
huggingface-hub
langdetect
marimo
matplotlib
numpy
openai
pandas
peft
python-dotenv
pyyaml
requests
ruff
sacrebleu
scikit-learn
seaborn
sentencepiece
tinker
tinker-cookbook
torch
tqdm
transformers
trl
